1-value work-around didn't work for images with non-float intensity values

+++ python/freesurfer/samseg/SamsegUtility.py
@@ -327,7 +327,7 @@ def writeImage(fileName, buffer, cropping, example):
 
 def logTransform(imageBuffers, mask):
 
-    logImageBuffers = imageBuffers.copy()
+    logImageBuffers = imageBuffers.copy().astype( 'float' )
     logImageBuffers[ logImageBuffers == 1 ] += 1e-5 # Voxels with zero values but inside the mask 
                                                     # should not be skipped in the C++ code!
     logImageBuffers[np.logical_not(mask), :] = 1
